feat: 게시물 추가 기능 개선 및 상태 관리 리팩토링
- addPost 함수를 addPostAPI로 이름 변경하여 API 호출 명확성 향상
- AddPostDialog 컴포넌트에서 props 제거 및 useAddPost 훅을 사용하여 상태 관리 개선
- PostsManagerPage에서 AddPostDialog의 props 간소화
