+++ src/features/posts/add-post/ui/AddPostDialog.tsx
@@ -9,10 +9,12 @@ import {
   Input,
   Textarea,
 } from '../../../../shared/ui';
-import { AddPostDialogProps } from '../model/type';
+import { useAddPost } from '../model/hooks';
 
-export const AddPostDialog = ({ newPost, onPostChange, onSubmit }: AddPostDialogProps) => {
+export const AddPostDialog = () => {
   const { isDialogOpen, closeDialog } = useDialogStore();
+  const { newPost, setNewPost, addPost } = useAddPost();
+
   return (
     <Dialog
       open={isDialogOpen(DIALOG_KEYS.ADD_POST)}
@@ -26,21 +28,21 @@ export const AddPostDialog = ({ newPost, onPostChange, onSubmit }: AddPostDialog
           <Input
             placeholder='제목'
             value={newPost.title}
-            onChange={(e) => onPostChange({ ...newPost, title: e.target.value })}
+            onChange={(e) => setNewPost({ ...newPost, title: e.target.value })}
           />
           <Textarea
             rows={30}
             placeholder='내용'
             value={newPost.body}
-            onChange={(e) => onPostChange({ ...newPost, body: e.target.value })}
+            onChange={(e) => setNewPost({ ...newPost, body: e.target.value })}
           />
           <Input
             type='number'
             placeholder='사용자 ID'
             value={newPost.userId}
-            onChange={(e) => onPostChange({ ...newPost, userId: Number(e.target.value) })}
+            onChange={(e) => setNewPost({ ...newPost, userId: Number(e.target.value) })}
           />
-          <Button onClick={onSubmit}>게시물 추가</Button>
+          <Button onClick={addPost}>게시물 추가</Button>
         </div>
       </DialogContent>
     </Dialog>
